pyproject.toml
[project]
name = "chembl-db-tools"
version = "0.1.0"
description = "Local tools for querying ChEMBL with LLMs"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "polars>=1.0.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
anthropic = [
  "anthropic>=0.32.0",
]
chemistry = [
  "rdkit-pypi>=2022.9.5",
]
local = [
  "torch>=2.3.0",
  "transformers>=4.43.0",
  "tokenizers>=0.19.0",
]

[dependency-groups]
dev = []
